DASH: prevent integer overflow
A integer overflow could happen in bandwidth computation, for example if
we have a DASH server and client on the same LAN.

+++ modules/stream_filter/dash/http/Chunk.cpp
@@ -72,7 +72,7 @@ void                Chunk::setUseByteRange  (bool value)
 {
     this->hasByteRange = value;
 }
-void                Chunk::setBitrate       (int bitrate)
+void                Chunk::setBitrate       (uint64_t bitrate)
 {
     this->bitrate = bitrate;
 }
