+++ modules/stream_filter/dash/mpd/IsoffMainManager.h
@@ -52,9 +52,9 @@ namespace dash
                 Period*                         getNextPeriod           (Period *period);
                 Representation*                 getBestRepresentation   (Period *period);
                 std::vector<Segment *>          getSegments             (const Representation *rep);
-                Representation*                 getRepresentation       (Period *period, int bitrate) const;
+                Representation*                 getRepresentation       (Period *period, uint64_t bitrate) const;
                 const MPD*                      getMPD                  () const;
-                Representation*                 getRepresentation       (Period *period, int bitrate,
+                Representation*                 getRepresentation       (Period *period, uint64_t bitrate,
                                                                          int width, int height) const;
 
             private:
